backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
django==3.0.5             # via -r requirements.in, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
requests==2.23.0          # via -r requirements.in
